requirements2.txt
torch==1.9.0
torchvision==0.10.0
numpy==1.21.2
streamlit==1.4.0
Pillow==8.3.1
scikit-image==0.18.3
scikit-learn==0.24.2
